[fix] model의 custom 타입들 선언 위치 수정
